feat: add mock data handling for contacts in development environment

+++ src/feature/contact/lib/contacts.ts
@@ -1,12 +1,20 @@
 "use server";
 
 import { isAuthenticatedGuard } from "@/feature/auth/lib/session";
+import {
+	getMockContactById,
+	mockContacts,
+} from "@/feature/contact/lib/mockContacts";
 import { prisma } from "@/lib/prisma";
 
 export const getAllContacts = async () => {
 	const authError = await isAuthenticatedGuard();
 	if (authError) {
 		return authError;
+	}
+
+	if (process.env.NODE_ENV === "development") {
+		return mockContacts;
 	}
 
 	return prisma.contact.findMany({
@@ -31,6 +39,10 @@ export const getContactById = async (id: string) => {
 		return authError;
 	}
 
+	if (process.env.NODE_ENV === "development") {
+		return getMockContactById(parseInt(id));
+	}
+
 	return prisma.contact.findUnique({
 		select: {
 			content: true,
@@ -51,6 +63,22 @@ export const deleteContact = async (id: string) => {
 		return authError;
 	}
 
+	if (process.env.NODE_ENV === "development") {
+		const mock = getMockContactById(parseInt(id));
+		if (!mock) {
+			throw new Error("Contact not found");
+		}
+		// delete mock contact from mockContacts
+		const index = mockContacts.findIndex(
+			(contact) => contact.id === parseInt(id),
+		);
+		if (index !== -1) {
+			mockContacts.splice(index, 1);
+		}
+
+		return mock;
+	}
+
 	return prisma.contact.delete({
 		where: { id: parseInt(id) },
 	});
@@ -60,6 +88,24 @@ export const deleteContacts = async (ids: string[]) => {
 	const authError = await isAuthenticatedGuard();
 	if (authError) {
 		return authError;
+	}
+
+	if (process.env.NODE_ENV === "development") {
+		// delete mock contacts from mockContacts
+		ids.forEach((id) => {
+			const mock = getMockContactById(parseInt(id));
+			if (mock) {
+				const index = mockContacts.findIndex(
+					(contact) => contact.id === parseInt(id),
+				);
+				if (index !== -1) {
+					mockContacts.splice(index, 1);
+				}
+			}
+		});
+		return {
+			count: ids.length,
+		};
 	}
 
 	return prisma.contact.deleteMany({
